requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
aiohttp>=3.8.0
pyyaml>=6.0
shapely>=2.0.0
